Add database support to flask_app lib, and add initial route
The both the app and the api make use of the same flask_app module. Eventually, that module should be refactored out of the services so that both services can inherit that shared module instead of managing their own copies of it. The same can be said of their Containerfiles, which currently happen to be exactly the same, and produce the same layers up until the differing project-level layers. Albeit minimal anyways, fortunately, this *shouldn't* produce redunant work if layer caching works as expected, so having them as seperate Containerfiles may be okay for now, despite them being the same for now